adding command to write dep graph json
Parsing the ndjson file turns out to take a really
long time (several minutes for a ~80 MB file). 

Therefore, similar to the other functions, it's worth
adding the ability to calculate them separately and write
the output to be re-used later. The result with a 20MB ndjson
output from bazel is a 2MB json file - so a 90% savings.